backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tax-filing-assistant"
version = "0.1.0"
description = "LLM-powered tax filing assistant with tool/function calling"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "pydantic>=2.7",
    "pydantic-settings>=2.4",
    "langgraph>=0.2.40",
    "langgraph-checkpoint-sqlite>=2.0",
    "langchain-core>=0.3",
    "langchain-openai>=0.2",
    "sqlmodel>=0.0.22",
    "pypdfium2>=4.30",
    "pdfplumber>=0.11",
    "pytesseract>=0.3.13",
    "pillow>=10.4",
    "python-multipart>=0.0.9",
    "sse-starlette>=2.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx>=0.27",
    "ruff>=0.6",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["../tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py311"
